requirements.txt
asgiref==3.5.2
colorama==0.4.5
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.6
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-import-export==2.8.0
django-js-asset==2.0.0
django-qr-code==3.1.0
django-resized==1.0.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
pydantic==1.9.1
pytz==2022.1
PyYAML==6.0
qrcode==7.3.1
segno==1.5.2
sqlparse==0.4.2
tablib==3.2.1
typing-extensions==4.3.0
tzdata==2022.1
xlrd==2.0.1
xlwt==1.3.0
